requirements.txt
fastapi
uvicorn[standard]
openai
speechrecognition
pyttsx3
pyaudio
pyautogui
python-dotenv
pycaw
comtypes
pvporcupine
mediapipe
opencv-python
pillow
pystray
psutil
websockets
requests
